fix: #1742 Fix invalid consumption cost example table markup
