+++ packages/marketplace/src/components/ui/developer-analytics/cost-explorer/cost-calculator/cost-calculator-form.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react'
 import { Form, Formik } from 'formik'
-import { EndpointsUsedRange } from './cost-calculator'
 import {
   Grid,
   GridItem,
@@ -12,10 +11,10 @@ import {
   isNumberOnly,
 } from '@reapit/elements'
 import styles from '@/styles/pages/developer-analytics.scss?mod'
+import { EndpointsUsedRange, endpointsUsedRange } from './use-foundation-cost-table'
 
 export type CostCalculatorFormProps = {
   initialValues: CostCalculatorFormValues
-  endpointsUsedRange: EndpointsUsedRange
   onSubmit: (values: CostCalculatorFormValues) => void
   onClear: () => void
 }
@@ -53,12 +52,7 @@ export const validate = (values: CostCalculatorFormValues): FormikErrors<CostCal
   return errors
 }
 
-const CostCalculatorForm: React.FC<CostCalculatorFormProps> = ({
-  initialValues,
-  endpointsUsedRange,
-  onSubmit,
-  onClear,
-}) => {
+const CostCalculatorForm: React.FC<CostCalculatorFormProps> = ({ initialValues, onSubmit, onClear }) => {
   return (
     <div className={styles.costCalculatorFormContainer}>
       <Formik initialValues={initialValues} validate={validate} onSubmit={onSubmit} enableReinitialize>
