+++ packages/marketplace/src/components/ui/developer-analytics/cost-explorer/cost-calculator/total-cost-table.tsx
@@ -1,104 +1,13 @@
 import * as React from 'react'
 import { FaCaretDown, FaCaretUp } from 'react-icons/fa'
 import { CostCalculatorFormValues } from './cost-calculator-form'
-import { EndpointsUsedRange, TierPrice } from './cost-calculator'
 import { formatCurrency, formatNumber } from '@/utils/number-formatter'
 import { Table, Grid, GridItem, H6, H4, Button } from '@reapit/elements'
+import useFoundationCostTable, { endpointsUsedRange } from './use-foundation-cost-table'
 import styles from '@/styles/pages/developer-analytics.scss?mod'
 
 export type TotalCostTableProps = {
   formValues: CostCalculatorFormValues
-  endpointsUsedRange: EndpointsUsedRange
-  foundationPricing: TierPrice
-}
-
-export type TableRow = {
-  numberOfApiCalls: number
-  costPerApiCall: number
-  totalCost: number
-}
-
-export type TotalCostTableData = {
-  tableData: TableRow[]
-  totalMonthlyCost: number
-}
-
-export const prepareData = (endpointsUsed: string, apiCalls: string, foundationPricing: TierPrice) => {
-  return (): TotalCostTableData => {
-    const bands: number[] = []
-    const bandCostPerCall: number[] = []
-    let totalMonthlyCost: number = 0
-
-    let remainingApiCalls = parseFloat(apiCalls)
-    const { priceRange, maxPrice } = foundationPricing[endpointsUsed]
-
-    for (const tierPriceLimit of priceRange) {
-      const { limit, price } = tierPriceLimit
-      bandCostPerCall.push(price)
-
-      if (remainingApiCalls > limit) {
-        bands.push(limit)
-        totalMonthlyCost += limit * price
-        remainingApiCalls -= limit
-      } else {
-        bands.push(remainingApiCalls)
-        totalMonthlyCost += remainingApiCalls * price
-        remainingApiCalls = 0
-        break
-      }
-    }
-
-    if (remainingApiCalls > 0) {
-      bands.push(remainingApiCalls)
-      bandCostPerCall.push(maxPrice)
-      totalMonthlyCost += maxPrice * remainingApiCalls
-    }
-
-    const costPerBand = bandCostPerCall.map((item, index) => {
-      return item * bands[index]
-    })
-
-    const tableData = bands.map((item, index) => {
-      return {
-        numberOfApiCalls: item,
-        costPerApiCall: bandCostPerCall[index],
-        totalCost: costPerBand[index],
-      }
-    })
-    return {
-      tableData,
-      totalMonthlyCost,
-    }
-  }
-}
-
-export const prepareTableColumns = (totalMonthlyCost: number) => {
-  return () => {
-    return [
-      {
-        Header: 'Number of API Calls',
-        accessor: row => {
-          return formatNumber(row.numberOfApiCalls)
-        },
-      },
-      {
-        Header: 'Cost Per API Call',
-        accessor: row => {
-          return formatCurrency(row.costPerApiCall, 6)
-        },
-        Footer: 'Estimated total monthly cost',
-      },
-      {
-        Header: 'Total Cost',
-        accessor: row => {
-          return formatCurrency(row.totalCost)
-        },
-        Footer: () => {
-          return formatCurrency(totalMonthlyCost)
-        },
-      },
-    ]
-  }
 }
 
 export const toggleShowTable = (isTableExpanded: boolean, setIsTableExpanded: (isTableExpanded: boolean) => void) => {
@@ -107,22 +16,14 @@ export const toggleShowTable = (isTableExpanded: boolean, setIsTableExpanded: (i
   }
 }
 
-const TotalCostTable: React.FC<TotalCostTableProps> = ({
-  formValues: { endpointsUsed, apiCalls },
-  endpointsUsedRange,
-  foundationPricing,
-}) => {
+const TotalCostTable: React.FC<TotalCostTableProps> = ({ formValues: { endpointsUsed, apiCalls } }) => {
   if (!endpointsUsed || !apiCalls) {
     return null
   }
 
   const [isTableExpanded, setIsTableExpanded] = React.useState(false)
-  const { tableData, totalMonthlyCost } = React.useMemo(prepareData(endpointsUsed, apiCalls, foundationPricing), [
-    endpointsUsed,
-    apiCalls,
-    foundationPricing,
-  ])
-  const tableColumns = React.useMemo(prepareTableColumns(totalMonthlyCost), [totalMonthlyCost])
+
+  const { tableColumns, tableData, totalMonthlyCost } = useFoundationCostTable(endpointsUsed, apiCalls)
 
   return (
     <div className="mt-5">
